refactor: separate out Ingredient and StorageIngredient
