+++ app/src/main/java/com/xffffff/wellfed/shoppingcart/ShoppingCartDB.java
@@ -3,18 +3,16 @@ package com.xffffff.wellfed.shoppingcart;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.xffffff.wellfed.common.DBConnection;
 import com.xffffff.wellfed.ingredient.Ingredient;
 import com.xffffff.wellfed.ingredient.IngredientDB;
-import com.xffffff.wellfed.ingredient.StorageIngredient;
-import com.xffffff.wellfed.ingredient.StorageIngredientDB;
+import com.xffffff.wellfed.storage.StorageIngredient;
+import com.xffffff.wellfed.storage.StorageIngredientDB;
 import com.xffffff.wellfed.mealplan.MealPlan;
 import com.xffffff.wellfed.mealplan.MealPlanDB;
 import com.xffffff.wellfed.recipe.Recipe;
-import com.xffffff.wellfed.recipe.RecipeDB;
 
 import java.util.ArrayList;
 import java.util.HashMap;
