+++ app/src/main/java/com/xffffff/wellfed/shoppingcart/ShoppingCartFragment.java
@@ -16,7 +16,7 @@ import com.xffffff.wellfed.R;
 import com.xffffff.wellfed.common.SearchInput;
 import com.xffffff.wellfed.common.SortingFragment;
 import com.xffffff.wellfed.ingredient.IngredientEditContract;
-import com.xffffff.wellfed.ingredient.StorageIngredient;
+import com.xffffff.wellfed.storage.StorageIngredient;
 
 import java.util.Arrays;
