+++ app/src/main/java/com/xffffff/wellfed/shoppingcart/ShoppingCartIngredientController.java
@@ -7,8 +7,8 @@ import com.google.firebase.firestore.Query;
 import com.xffffff.wellfed.common.DBConnection;
 import com.xffffff.wellfed.common.DateUtil;
 import com.xffffff.wellfed.ingredient.Ingredient;
-import com.xffffff.wellfed.ingredient.StorageIngredient;
-import com.xffffff.wellfed.ingredient.StorageIngredientDB;
+import com.xffffff.wellfed.storage.StorageIngredient;
+import com.xffffff.wellfed.storage.StorageIngredientDB;
 import com.xffffff.wellfed.mealplan.MealPlan;
 import com.xffffff.wellfed.mealplan.MealPlanController;
 import com.xffffff.wellfed.mealplan.MealPlanDB;
