+++ app/src/main/java/com/xffffff/wellfed/storage/StorageIngredient.java
@@ -1,5 +1,7 @@
-package com.xffffff.wellfed.ingredient;
+package com.xffffff.wellfed.storage;
 
+
+import com.xffffff.wellfed.ingredient.Ingredient;
 
 import java.util.ArrayList;
 import java.util.Date;
