+++ app/src/test/java/com/xffffff/wellfed/StorageIngredientTest.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-import com.xffffff.wellfed.ingredient.StorageIngredient;
+import com.xffffff.wellfed.storage.StorageIngredient;
 
 import org.junit.Test;
